add icon emojis
