+++ src/containers/SideBar/Miner/components/ModeSelect.tsx
@@ -4,8 +4,11 @@ import { useAppStatusStore } from '@app/store/useAppStatusStore.ts';
 import { useChangeMiningMode } from '@app/hooks/mining/useMiningControls';
 import { useTranslation } from 'react-i18next';
 import { Typography } from '@app/components/elements/Typography.tsx';
-import { Select } from '@app/components/elements/Select.tsx';
+import { Select } from '@app/components/elements/inputs/Select.tsx';
 import { useMiningStore } from '@app/store/useMiningStore.ts';
+
+import eco from '@app/assets/icons/emoji/eco.png';
+import fire from '@app/assets/icons/emoji/fire.png';
 
 function ModeSelect() {
     const { t } = useTranslation('common', { useSuspense: false });
@@ -27,8 +30,8 @@ function ModeSelect() {
                 onChange={handleChange}
                 selectedValue={mode}
                 options={[
-                    { label: '♻️ ECO', value: 'Eco', selectedLabel: 'ECO ♻️' },
-                    { label: '🔥 Ludicrous', value: 'Ludicrous', selectedLabel: 'Ludicrous 🔥' },
+                    { label: 'ECO', value: 'Eco', iconSrc: eco },
+                    { label: 'Ludicrous', value: 'Ludicrous', iconSrc: fire },
                 ]}
             />
         </TileItem>
